Do not attempt to fall back to ANetAstrometryTask
In DM-8832, the simple try/except caused a misleading error: an error in the
config for AstrometryTask meant that it threw when we tried to load it, but
the except caught this and converted it into an attempt to configure
ANetAstrometryTask instead. Wasted time trying to figure out why the latter
was broken, when, in fact, we should never have been touching it at all.

Per DM-8209, the attempt to fall back is doomed to failure anyway: the
astrometry task cannot have been retargeted to ANetAstrometryTask due to the
order in which overrides are applied. It can therefore safely be removed.